fix: SpeechAPI run main scene
