+++ Assets/Scripts/Recognization.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.EventSystems;
 
 public class Recognization : MonoBehaviour {
 
@@ -13,17 +14,26 @@ public class Recognization : MonoBehaviour {
 
   // ボタンをクリックした時の処理
   public void OnClick() {
-    Debug.Log("Button click!");
+    if (!IsPointerOverUIObject())
+    {
+      Debug.Log("Button click!");
+    }
   }
 
   public void PushDown() {
-    push = true;
-    Debug.Log("Button push!");
+    if (Input.touchCount > 0 && !IsPointerOverUIObject())
+    {
+      push = true;
+      Debug.Log("Button push!");
+    }
   }
 
   public void PushUp() {
-    push = false;
-    Debug.Log("Button release!");
+    if (Input.touchCount > 0 && !IsPointerOverUIObject())
+    {
+      push = false;
+      Debug.Log("Button release!");
+    }
   }
 	
 	// Update is called once per frame
@@ -34,4 +44,14 @@ public class Recognization : MonoBehaviour {
       SpeechAPI.SpeechAPI.StopRecord();
     }
 	}
+
+  private bool IsPointerOverUIObject()
+  {
+    PointerEventData eventDataCurrentPosition = new PointerEventData(EventSystem.current);
+    eventDataCurrentPosition.position = new Vector2(Input.mousePosition.x, Input.mousePosition.y);
+    List<RaycastResult> results = new List<RaycastResult>();
+    EventSystem.current.RaycastAll(eventDataCurrentPosition, results);
+    return results.Count > 0;
+  }
+
 }
